Added version 3 that uses MQTT to store the time on the clock face on each movement
This version can find out, with an MQTT subscription at power up, what the time was on the clock face when power was removed. This means that we don't have to write to flash memory every 30 seconds, which would wear out the flash too quickly, just publish to an MQTT server every 30 seconds.