Update WallClockV3.toit
Fixed bug that added 30secs when restarting. I was pulsing at restart because I was missing a slot, but now that the clock stores the time of the clock face, it can just catch up, regardless of how many pulses it might miss at restart, so does not need an unrecorded pulse any more.